+++ problemes/probleme2xx/probleme210.cpp
@@ -17,7 +17,7 @@ ENREGISTRER_PROBLEME(210, "Obtuse Angled Triangles") {
 
     nombre k = r / 4;
     nombre RR = k * k / 2;
-    auto R = static_cast<nombre>(sqrtl(RR));
+    auto R = static_cast<nombre>(std::sqrt(static_cast<long double>(RR)));
 
     nombre compteur = 0;
     for (nombre x = k / 2, y = k / 2; x; --x) {
