+++ problemes/probleme2xx/probleme222.cpp
@@ -32,7 +32,7 @@ ENREGISTRER_PROBLEME(222, "Sphere Packing") {
     const nombre taille = 21;
     std::deque<long double> rayons;
     for (size_t n = 0; n < taille; ++n)
-        rayons.push_back((30.0L + n) / R);
+        rayons.push_back((30.0L + static_cast<long double>(n)) / R);
 
     std::random_device rd;
     std::mt19937 g(rd());
