+++ problemes/probleme2xx/probleme292.cpp
@@ -11,8 +11,8 @@ namespace {
     typedef std::tuple<short, short, short> triplet;
 
     void generer_triplets(const short longueur, std::vector<triplet> &triplets) {
-        for (short dx = -longueur; dx <= longueur; ++dx)
-            for (short dy = -longueur; dy <= longueur; ++dy) {
+        for (short dx = static_cast<short>(-longueur); dx <= longueur; ++dx)
+            for (short dy = static_cast<short>(-longueur); dy <= longueur; ++dy) {
                 if (auto racine = racine::carre_parfait(dx * dx + dy * dy)) {
                     if (racine.value() < longueur)
                         triplets.emplace_back(dx, dy, racine.value());
